requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-http-exceptions==1.4.0
django-modern-rpc==1.0.1
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
jsonrpcserver==5.0.9
jsonschema==4.17.3
OSlash==0.6.3
packaging==23.0
pluggy==1.0.0
py==1.11.0
pyrsistent==0.19.3
PyYAML==6.0
requests==2.28.2
simplejson==3.18.3
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
